% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.marima.R
\name{predict.marima}
\alias{predict.marima}
\title{predict.marima}
\usage{
\method{predict}{marima}(
  fit,
  newdata = NULL,
  nstart = 1,
  nstep = 1,
  dif.poly = NULL,
  check = TRUE
)
}
\description{
Forecasting of (multivariate) time series of
using marima type model.
}
